NAV MESH!!!!
Okay lots done here, removed the old Node class and replaced it with
MapNodes and NaveMeshNodes. NaveMeshNodes use MapNodes to build the mesh
off using four of them to make up all the cornoersand uses those then to
figure out who the NavMeshNode is borardering to make a connection with.
the Astar now uses the NavMeshNodes and welp now everything uses the
NaveMeshNodes sooo hurray!!!!